+++ decomp/SDK/include/dolphin/os/OSSerial.h
@@ -28,36 +28,36 @@
 
 struct SIControl
 {
-    long          chan;
-    unsigned long poll;
-    unsigned long inputBytes;
+    s32          chan;
+    u32 poll;
+    u32 inputBytes;
     void*         input;
-    void          (*callback)(long, unsigned long, struct OSContext*);
+    void          (*callback)(s32, u32, struct OSContext*);
 };
 
 struct SIPacket
 {
-    long          chan;
+    s32          chan;
     void*         output;
-    unsigned long outputBytes;
+    u32 outputBytes;
     void*         input;
-    unsigned long inputBytes;
-    void          (*callback)(long, unsigned long, struct OSContext*);
-    long long     time;
+    u32 inputBytes;
+    void          (*callback)(s32, u32, struct OSContext*);
+    s64     time;
 };
 
 int           SIBusy();
 void          SIInit();
-unsigned long SISync();
-unsigned long SIGetStatus();
-void          SISetCommand(long chan, unsigned long command);
-unsigned long SIGetCommand(long chan);
+u32 SISync();
+u32 SIGetStatus();
+void          SISetCommand(s32 chan, u32 command);
+u32 SIGetCommand(s32 chan);
 void          SITransferCommands();
-unsigned long SISetXY(unsigned long x, unsigned long y);
-unsigned long SIEnablePolling(unsigned long poll);
-unsigned long SIDisablePolling(unsigned long poll);
-void          SIGetResponse(long chan, void* data);
+u32 SISetXY(u32 x, u32 y);
+u32 SIEnablePolling(u32 poll);
+u32 SIDisablePolling(u32 poll);
+void          SIGetResponse(s32 chan, void* data);
 int           SITransfer(s32 chan, void* output, u32 outputBytes, void* input, u32 inputBytes,
-                         void (*callback)(long, u32, struct OSContext*), s64 time);
+                         void (*callback)(s32, u32, struct OSContext*), s64 time);
 
 #endif // _DOLPHIN_OSSERIAL_H
