+++ decomp/SDK/src/dolphin/gx/GXAttr.c
@@ -136,10 +136,10 @@ __GXSetVCD(void)
     static u8     tbl1[] = { 0, 4, 1, 2 };
     static u8     tbl2[] = { 0, 8, 1, 2 };
     static u8     tbl3[] = { 0, 12, 1, 2 };
-    unsigned long vlm;
-    unsigned long b;
-    unsigned long vl;
-    unsigned long vh;
+    u32 vlm;
+    u32 b;
+    u32 vl;
+    u32 vh;
 
     GX_WRITE_SOME_REG4(8, 0x50, __GXData->vcdLo, -12);
     GX_WRITE_SOME_REG4(8, 0x60, __GXData->vcdHi, -12);
